Converting to use Eloquent and Query Builder
